+++ plugins/vcs-core/core/source_gen/jetbrains/mps/vcs/diff/merge/MergeConflictsBuilder.java
@@ -237,7 +237,7 @@ public class MergeConflictsBuilder {
         return mvc;
       }
     }, ModelVersionChange.class);
-    if ((int) MapSequence.fromMap(arranged._0()).count() == 1 && (int) MapSequence.fromMap(arranged._1()).count() == 1) {
+    if (MapSequence.fromMap(arranged._0()).count() == 1 && MapSequence.fromMap(arranged._1()).count() == 1) {
       addPossibleConflict(SetSequence.fromSet(MapSequence.fromMap(arranged._0()).keySet()).first(), SetSequence.fromSet(MapSequence.fromMap(arranged._1()).keySet()).first());
     }
   }
